refactor(db): move note link sync writes to repository (DB-REPOSITORY-PILOT-NEXT-C2)
- 新增 noteLinksRepository.replaceLinksForSource() 方法
- 在 types.ts 中添加 NoteLinkEntry 类型
- 修改 noteLinks.ts 调用 Repository
- 保持现有事务边界：DELETE 不在事务内，INSERT 在事务内
